multiple items ship with this commit: - builder for network - started writing a BasicScenarioLoader currently supporting only Network loading - merged functionality of BasicScenarioLoader and ScenarioLoader - merged functionality of Scenario and BasicScenario - Comparator for households - convenience constructor for HouseholdsWriter - removed Coord check in BasicNodeImpl according to interface definition of Node, i.e. the Coord must not exist, removed very old deprecated todo - made NetworkFactory implement NetworkBuilder interface - added builder to Network containers and set other create methods to deprecated - added util method in MGC - households creator and chartdata writer for income studies + some utils (playground only) - changes in income scoring function and testcase (playground only)
git-svn-id: file:///srv/local/data/cleanRepo/trunk@7227 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/org/matsim/core/network/algorithms/NetworkWriteAsTable.java
@@ -100,7 +100,7 @@ public class NetworkWriteAsTable {
 	public void run(Network network) {
 		System.out.println("    running " + this.getClass().getName() + " algorithm...");
 
-		int capperiod = network.getCapacityPeriod();
+		double capperiod = network.getCapacityPeriod();
 		capperiod = capperiod / 3600;
 		System.out.println("      capperiod = " + capperiod);
