+++ matsim/src/playground/balmermi/algos/NetworkWriteVolumesAsTable.java
@@ -81,7 +81,7 @@ public class NetworkWriteVolumesAsTable {
 	public void run(Network network) {
 		System.out.println("    running " + this.getClass().getName() + " algorithm...");
 
-		int capperiod = network.getCapacityPeriod();
+		double capperiod = network.getCapacityPeriod();
 		capperiod = capperiod / 3600;
 		System.out.println("      capperiod = " + capperiod);
